campusx/Story_Generator_Project/requirements.txt
streamlit
python-dotenv  # for working on load_dotenv
google-genai
google-generativeai
Pillow
gTTS